+++ packages/projectlint/scripts/test-codemod.js
@@ -1,4 +1,4 @@
-const ProjectLint = require('../dist').default
+const ProjectLint = require('../dist').default;
 
 const args = process.argv.slice(2);
 const [cwd] = args;
@@ -11,4 +11,4 @@ const projectLint = new ProjectLint({ cwd, transforms: { snapshot: 1, picture: 1
 
   const fixResult = await projectLint.fix();
   console.log('projectLint fix result: ', fixResult);
-})()
+})();