+++ packages/eslint-config/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - fix: set `dot-notation` off
 - fix: set `prefer-const` off
+- fix: set `react-hooks/rules-of-hooks` warn
+- fix: set `react/jsx-uses-react` warn
+- fix: set `react/jsx-uses-vars` warn
 
 ## 0.1.0